1. Отделил базовые классы от классов-тестов; 2. Переписал тесты валидатора чисел на fluent assertions.
