+++ cs/HomeExercises/TsarRegistry.cs
@@ -1,0 +1,13 @@
+namespace HomeExercises
+{
+    public class TsarRegistry
+    {
+		public static Person GetCurrentTsar()
+		{
+			return new Person(
+				"Ivan IV The Terrible", 54, 170, 70,
+				new Person("Vasili III of Russia", 28, 170, 60, null));
+		}
+        
+    }	
+}
